% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_stick.R
\name{dp_stick}
\alias{dp_stick}
\title{dp_stick}
\usage{
dp_stick(n = 1, M = 1, fun, ...)
}
\arguments{
\item{n}{Number of partitions to make.}

\item{M}{Precision of the baseline distribution.}

\item{fun}{The distribution to sample from.}

\item{...}{Parameters for the distribution.}
}
\value{
Data frame with mass points and the corresponding weights, an approximation to the Dirichlet process.
}
\description{
This function generates an approximation of the Dirichlet Process via the stick-breaking construction. Notice that the probabilities won't sum 1.
}
\examples{
dp_stick(n = 10, M = 1, rnorm, mean = 0, sd = 1)
dp_stick(n = 100, M = 7, rgamma, shape = 2, rate = 2)
}
