% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_urn.R
\name{dp_urn}
\alias{dp_urn}
\title{dp_urn}
\usage{
dp_urn(n = 1, M = 1, fun, ...)
}
\arguments{
\item{n}{Number of times the scheme is set to run.}

\item{M}{Precision of the baseline density.}

\item{fun}{Density that models the data.}

\item{...}{Parameters for the density.}
}
\value{
A data frame containing the mass points with their weights, an approximation to a Dirichlet process.
}
\description{
This function approximates a Dirichlet Process by a generalization of the
Polya urn scheme.
}
\examples{
dp_urn(n = 10, M = 3, rnorm, mean = 0, sd = 1)
dp_urn(n = 10, M = 3, rgamma, shape = 2, rate = 2)
}
